requirements.txt
asgiref==3.4.1
boto3==1.18.12
botocore==1.21.12
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.46.0
django-cors-headers==3.7.0
django-environ==0.7.0
django-heroku==0.3.1
django-utils-six==2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.11.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
loguru==0.5.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
paypalrestsdk==1.13.1
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.2
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
uuid==1.30
validators==0.18.2
whitenoise==5.3.0
